+++ src/components/posts/posts/Tabs/share-tab.jsx
@@ -1,9 +1,12 @@
-import * as S from '../../../../styles/share-tab';
+import * as S from '../../../../styles/posts/posts/Tabs/share-tab';
 
 const ShareTab = () => {
     return(
         <div>
-
+            <div>공유하기</div>
+            <div>
+                
+            </div>
         </div>
     );
 }
