+++ src/styles/posts/posts/Tabs/friend-card.js
