+++ src/styles/posts/posts/Tabs/friends-tab.js
