+++ src/styles/posts/posts/Tabs/share-tab.js
